Move HelperText enum to common form style
